+++ src/components/show/Show.js
@@ -48,9 +48,14 @@ export default function Show() {
     const onPress= async (e)=>{
         e.preventDefault();
         if (defname === "Remove Favorite"){
-            await axios.delete(`http://localhost:8080/user/favshows/delete/${localStorage.getItem('userid')}/${showid}`);
-            navigate(`/user/${localStorage.getItem('userid')}`)
-        }else if (defname === "Set Favorite"){
+            const response = await axios.get(`http://localhost:8080/user/favshows/${localStorage.getItem('userid')}`);
+            if (response.data.length === 1){
+                alert("You need at least 1 favorite")
+            }else {
+                await axios.delete(`http://localhost:8080/user/favshows/delete/${localStorage.getItem('userid')}/${showid}`);
+                navigate(`/user/${localStorage.getItem('userid')}`)
+            }
+            }else if (defname === "Set Favorite"){
             await axios.put(`http://localhost:8080/user/favshows/${localStorage.getItem('userid')}/${showid}`);
             navigate(`/user/${localStorage.getItem('userid')}`);
         }
@@ -97,6 +102,10 @@ export default function Show() {
                                     <a target={"_blank"} href={show.imdb_url}>
                                         Website
                                     </a>
+                                </li>
+                                <li className="list-group-item ">
+                                    <b>Favorite Counter: </b>
+                                    {show.favorite_count}
                                 </li>
                                 <li className="list-group-item ">
                                     <button type="button" onClick={onPress}  className={`btn ${btn}`}>{defname}</button>
